+++ zest/releaser/utils.py
@@ -597,13 +597,6 @@ def run_hooks(setup_cfg, which_releaser, when, data):
         hook_names = config.get('zest.releaser', hook_group).split()
         hooks = []
 
-        # Hooks that can't be imported are skipped, unless the
-        # ``skip_missing_hooks`` option is set to False.
-        skip_missing = True
-        if config.has_option('zest.releaser', 'skip_missing_hooks'):
-            skip_missing = config.getboolean('zest.releaser',
-                                             'skip_missing_hooks')
-
         # The following code is adapted from the 'packaging' package being
         # developed for Python's stdlib:
 
@@ -623,13 +616,9 @@ def run_hooks(setup_cfg, which_releaser, when, data):
 
         try:
             for hook_name in hook_names:
-                try:
-                    hooks.append(resolve_name(hook_name))
-                except ImportError as e:
-                    if not skip_missing:
-                        raise
-                    logger.warning('cannot find %s hook: %s; skipping...',
-                                   hook_name, e.args[0])
+                # Resolve the hook or fail with ImportError.
+                hooks.append(resolve_name(hook_name))
+
             for hook in hooks:
                 hook(data)
         finally: